+++ src/retractions/retractions.jl
@@ -227,7 +227,7 @@ end
 
 Compute the Cayley retraction of `B`.
 
-This is equivalent to the method of [`cayley`](@ref) for [StiefelLieAlgHorMatrix](@ref).
+This is equivalent to the method of [`cayley`](@ref) for [`StiefelLieAlgHorMatrix`](@ref).
 
 See [`cayley(::StiefelLieAlgHorMatrix)`](@ref).
 """